Added categories to posts
