+++ package.json
@@ -30,11 +30,12 @@
     "classnames": "^2.5.1",
     "firebase": "^10.14.0",
     "firebase-admin": "^11.11.1",
-    "next": "^14.2.11",
+    "next": "^14.2.14",
     "next-auth": "^5.0.0-beta.20",
     "react": "^18",
     "react-dom": "^18",
     "react-icons": "^5.0.1",
+    "request": "^2.88.2",
     "server-only": "^0.0.1",
     "tiny-invariant": "^1.3.3"
   },
