--- src/app/publicar/categories.ts
@@ -1,11 +1,0 @@
-export const categories = [
-    { id: "1", name: "Ensaladas" },
-    { id: "2", name: "Vegeratariano" },
-    { id: "3", name: "A la parrilla" },
-    { id: "4", name: "Pescados y mariscos" },
-    { id: "5", name: "Bajos en calorias" },
-    { id: "6", name: "Smoothies y Jugos Verdes" },
-    { id: "7", name: "Wraps integraless" },
-    { id: "8", name: "Sin gluten" },
-    { id: "9", name: "Balanceada" }
-  ];
